add order list, keep login feature

+++ frontend/src/pages/LoginPage.js
@@ -16,6 +16,7 @@ const LoginPage = () => {
     e.preventDefault()
     const username = e.target.username.value
     const password = e.target.password.value
+    const keepLogin = document.querySelector('.save-account').checked;
 
     const response = await login(username, password)
 
@@ -24,7 +25,12 @@ const LoginPage = () => {
     } else if (response === "password not match") {
       errorToast("Mật khẩu không đúng")
     } else {
-      localStorage.setItem('token', JSON.stringify(response))
+      if (keepLogin) {
+        localStorage.setItem('token', JSON.stringify(response))
+        localStorage.setItem('keepLogin', JSON.stringify(true))
+      } else {
+        sessionStorage.setItem('token', JSON.stringify(response))
+      }
       cartContext.setCart(await getCart(getUserId()))
       navigate('/profile/')
     }
@@ -41,7 +47,7 @@ const LoginPage = () => {
           <input type="password" placeholder="Mật khẩu" name="password" required/>
         </div>
         <div class="remember-password">
-          <input type="checkbox"/>
+          <input type="checkbox" className='save-account' />
           <h3>Lưu tài khoản</h3>
         </div>
         <div class="input-box button">
